deletions and movals
