+++ tests/test_harness.py
@@ -1,17 +1,13 @@
-
 import unittest
 import os
-from active_tools import make_an_array_instance_active
-from dummy_data import make_test_ncdata
+from activestorage.dummy_data import make_test_ncdata
 from netCDF4 import Dataset
 import numpy as np
 from numcodecs.compat import ensure_ndarray
 from zarr.indexing import (
     OrthogonalIndexer,
-    check_fields,
 )
-
-import netcdf_to_zarr as nz
+from activestorage import netcdf_to_zarr as nz
 
 class Active:
     """ 
@@ -79,96 +75,51 @@ class Active:
                 self.zds = ds
             return self
         else:
-            return self.__doing_it_ourselves2(*args)
+            return self.__get_selection(*args)
 
-    def __vanilla_zarr(self, *args):
-        return self.zds.__getitem__(*args)
 
-    def __doing_it_ourselves(self,*args):
-        """ This is the first version unpacking stuff ourselves directly.
-        Doesn't quite get things in the right order, whether we read the 
-        data or whether zarr does iet."""
-
-        data_selection, chunk_info, chunk_coords = self.zds.get_orthogonal_selection(*args,
-                                                 out=None, fields=None)
-
-        chunks, chunk_sel, PCI= chunk_info[0]
-
-        # get offsets and sizes from PCI
-        offsets = []
-        sizes = []
-        for offset, size, _ in list(PCI):
-            offsets.append(offset)
-            sizes.append(size)
-
-        # get chunks info from chunk store
-        chunk_store = self.zds.chunk_store
-        chunk_coords_formatted = []
-        for ch_coord in chunk_coords:
-            new_key = "data/" + ".".join([str(ch_coord[0]),
-                                        str(ch_coord[1]),
-                                        str(ch_coord[2])])
-            chunk_coords_formatted.append(new_key)
-
-        # decode bytes from chunks
-        # this is vanilla zarr
-        chunks_with_data = [self.zds._decode_chunk(chunk_store[k]) for k in chunk_coords_formatted]
-
-        # this is going to the file ourselves directly
-        chunks_our_way = [self._decode_chunk(k) for k in chunk_coords_formatted]
-
-        print('Zarr',chunks_with_data)
-        print('Us',chunks_our_way)        
-
-        chunks_dict = {}
-        for (i, k), f in zip(enumerate(chunk_coords_formatted), chunk_coords):
-            flat_decoded = np.ndarray.flatten(self.zds._decode_chunk(chunk_store[k]))
-            selection_in_chunk = []
-            # NB: very important to remember that each start in "offsets" is to be
-            # used for each chunk; it's not one start from "offsets" is per chunk
-            for j, k in zip(offsets, sizes):
-                partial_data = flat_decoded[j:j+k]
-                selection_in_chunk.extend(partial_data)
-            chunks_dict[f] = selection_in_chunk
-
-        selection = []
-        for _, v in chunks_dict.items():
-            selection.extend(v)
-
-        return np.array(selection)
-
-    def __doing_it_ourselves2(self, *args):
-
-        # conflating _get_orthogonal_selection ...
-
-        # we probably want to get the results of needing ot use indexer and pass that down
-        # as I suspect this might not be something to push down into a real active layer.
+    def __get_selection(self, *args):
+        """ 
+        First we need to convert the selection into chunk coordinates,
+        steps etc, via the Zarr machinery, then we get everything else we can
+        from zarr and friends and use simple dictionaries and tupes, then
+        we can go to the storage layer with no zarr.
+        """
+        if self.zds._compressor:
+            raise ValueError("No active support for compression as yet")
+        if self.zds._filters:
+            raise ValueError("No active support for filters as yet")
+        
         indexer = OrthogonalIndexer(*args, self.zds)
-
-        # ... with the start of as_get_selection
-        out_dtype = self.zds._dtype
         out_shape = indexer.shape
+        out_dtype = self.zds._dtype
+        stripped_indexer = [(a, b, c) for a,b,c in indexer]
+        drop_axes = indexer.drop_axes  # not sure what this does and why, yet.
+
+        # yes this next line is bordering on voodoo ... 
+        fsref = self.zds.chunk_store._mutable_mapping.fs.references
+
+        return self.__from_storage(stripped_indexer, drop_axes, out_shape, out_dtype, fsref)
+
+    def __from_storage(self, stripped_indexer, drop_axes, out_shape, out_dtype, fsref):
+ 
         out = np.empty(out_shape, dtype=out_dtype, order=self.zds._order)
 
-        for chunk_coords, chunk_selection, out_selection in indexer:
-            self._process_chunk(chunk_coords,chunk_selection, out, out_selection,
-                                    drop_axes=indexer.drop_axes) 
+        for chunk_coords, chunk_selection, out_selection in stripped_indexer:
+            self._process_chunk(fsref, chunk_coords,chunk_selection, out, out_selection,
+                                    drop_axes=drop_axes) 
         
         if self.method is not None:
             return self.method(out)
         else:
             return out
 
-    def _process_chunk(self, chunk_coords, chunk_selection, out, out_selection,
+    def _process_chunk(self, fsref, chunk_coords, chunk_selection, out, out_selection,
                        drop_axes=None):
         """Obtain part or whole of a chunk by taking binary data from storage and filling output array"""
-        if self.zds._compressor:
-            raise ValueError("No active support for compression as yet")
-        if self.zds._filters:
-            raise ValueError("No active support for filters as yet")
-        
+      
         key = "data/" + ".".join([str(c) for c in chunk_coords])
-        chunk = self._decode_chunk(key)
+        chunk = self._decode_chunk(fsref, key)
         tmp = chunk[chunk_selection]
         if drop_axes:
             tmp = np.squeeze(tmp, axis=drop_axes)
@@ -176,12 +127,11 @@ class Active:
         # store selected data in output
         out[out_selection] = tmp
 
-    def _decode_chunk(self, key):
+    def _decode_chunk(self,fsref, key):
         """ We do our own read of chunks and decoding etc """
         
-        # yes this next line is bordering on voodoo ... 
-        myfsref = self.zds.chunk_store._mutable_mapping.fs.references
-        rfile, offset, size = tuple(myfsref[key])
+        
+        rfile, offset, size = tuple(fsref[key])
         #fIXME: for the moment, open the file every time ... we might want to do that, or not
         with open(rfile,'rb') as open_file:
             # get the data
@@ -260,39 +210,6 @@ class TestActive(unittest.TestCase):
         var = active['data']
         result2 = var[0:2,4:6,7:9]
         self.assertEqual(mean_result, result2)
-       
-
-    def test_zarr_hijack(self):
-        """ 
-        Test the hijacking of Zarr. 
-        """
-        data_file = self.testfile
-        varname = "test_bizarre"
-        selection = (slice(0, 2, 1), slice(4, 6, 1), slice(7, 9, 1))
-
-        # get slicing info
-        (ds, master_chunks, chunk_selection,
-         offsets, sizes, selected_chunks, chunks_data_dict) = \
-            nz.slice_offset_size(data_file, varname, selection)
-
-        # sanity checks
-        assert master_chunks == (3, 3, 1)
-        assert len(offsets) == 2
-        assert len(sizes) == len(offsets)
-        assert offsets == [1, 4]  # these are numbers of elements
-        # and are found in each chunk ie 1st and 4th element start data in each chunk
-        assert sizes == [2, 2]  # these are numbers of elements
-        # and are found in each chunk ie 1st+2 and 4th+2 span selected data in each chunk
-        assert selected_chunks == [(0, 1, 7), (0, 1, 8)]  # chunks coords of the chunks containing selected data
-        assert selected_chunks == list(chunks_data_dict.keys())
-        expected_selection = np.sort([740.,840.,750.,850.,741.,841.,751.,851.])
-        selection = []
-        for _, v in chunks_data_dict.items():
-            selection.extend(v)
-        print(f"UNITTEST: expected data selection: {expected_selection}")
-        print(f"UNITTEST: obtained data selection: {np.array(selection)}")
-        assert np.array_equal(np.array(expected_selection), np.sort(selection))
-
 
 if __name__=="__main__":
     unittest.main()
